app/api/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.3.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
datasets==4.1.1
dill==0.4.0
fastapi==0.117.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
langcodes==3.5.0
langdetect==1.0.9
language-data==1.3.0
libclang==18.1.1
lxml==6.0.2
marisa-trie==1.3.1
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml-dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
namex==0.1.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
pluggy==1.6.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.5
pypdf2==3.0.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf-keras==2.20.1
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
